+++ app/build.gradle
@@ -1,16 +1,19 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
+    id ("kotlin-kapt")
+    id("kotlin-parcelize")
+    id("androidx.navigation.safeargs")
 }
 
 android {
     namespace 'com.dprieto.dragonballapp'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.dprieto.dragonballapp"
         minSdk 21
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -43,6 +46,11 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
     implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
+
+    //Images
+    implementation("io.coil-kt:coil:2.2.2")
+
+
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
