+++ app/src/main/AndroidManifest.xml
@@ -12,11 +12,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.DragonBallApp"
         tools:targetApi="31">
+
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.DragonBallApp.NoActionBar">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
