+++ app/src/main/java/com/dprieto/dragonballapp/SplashFragment.kt
@@ -1,0 +1,62 @@
+package com.dprieto.dragonballapp
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
+import com.dprieto.dragonballapp.databinding.FragmentSplashBinding
+import kotlinx.coroutines.delay
+
+
+class SplashFragment : Fragment() {
+
+    private var _binding: FragmentSplashBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        _binding = FragmentSplashBinding.inflate(inflater, container, false)
+        return binding.root
+
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        lifecycleScope.launchWhenCreated {
+            goToLoginFragment()
+        }
+
+    }
+
+    private suspend fun goToLoginFragment() {
+        delay(3000)
+        findNavController()
+            .navigate(SplashFragmentDirections.actionSplashFragmentToLoginFragment())
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    override fun onResume() {
+        super.onResume()
+        (activity as AppCompatActivity?)?.supportActionBar?.hide()
+    }
+
+    override fun onStop() {
+        super.onStop()
+        (activity as AppCompatActivity?)?.supportActionBar?.show()
+    }
+}
